requirements.txt
django==1.10
requests
django-environ
python-crontab==2.2.8